+++ packages/frontend/src/styles/z-index.ts
@@ -1,6 +1,7 @@
 export enum ZIndex {
+  ControlBar = 900,
   Header = 1000,
-  HamburgerMenu = 1100,
   Bar = Header,
+  HamburgerMenu = 1100,
   FilterModal = 1200,
 }